🚚 api: remove route to certification/enrolment context
