+++ api/src/certification/enrolment/application/enrolment-route.js
@@ -30,6 +30,29 @@ const register = async function (server) {
         tags: ['api', 'sessions', 'students'],
       },
     },
+    {
+      method: 'GET',
+      path: '/api/sessions/{id}/candidates-import-sheet',
+      config: {
+        validate: {
+          params: Joi.object({
+            id: identifiersType.sessionId,
+          }),
+        },
+        pre: [
+          {
+            method: authorization.verifySessionAuthorization,
+            assign: 'authorizationCheck',
+          },
+        ],
+        handler: enrolmentController.getCandidatesImportSheet,
+        tags: ['api', 'sessions'],
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs appartenant à un centre de certification ayant créé la session**\n' +
+            "- Cette route permet de télécharger le template d'import des candidats d'une certification au format ods",
+        ],
+      },
+    },
   ]);
 };
