--- api/tests/acceptance/application/session/session-controller-get-candidates-import-sheet_test.js
@@ -1,49 +1,0 @@
-import {
-  createServer,
-  databaseBuilder,
-  expect,
-  generateValidRequestAuthorizationHeader,
-} from '../../../test-helper.js';
-
-describe('Acceptance | Controller | session-controller-get-candidates-import-sheet', function () {
-  let server;
-
-  beforeEach(async function () {
-    server = await createServer();
-  });
-
-  describe('GET /api/sessions/{id}/candidates-import-sheet', function () {
-    it('should respond with a 200 when session can be found', async function () {
-      // given
-      const user = databaseBuilder.factory.buildUser();
-      const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
-      databaseBuilder.factory.buildCertificationCenterMembership({ userId: user.id, certificationCenterId });
-
-      const otherUserId = databaseBuilder.factory.buildUser().id;
-      const otherCertificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
-      databaseBuilder.factory.buildCertificationCenterMembership({
-        userId: otherUserId,
-        certificationCenterId: otherCertificationCenterId,
-      });
-
-      const sessionIdAllowed = databaseBuilder.factory.buildSession({ certificationCenterId }).id;
-
-      await databaseBuilder.commit();
-
-      // when
-      const options = {
-        method: 'GET',
-        url: `/api/sessions/${sessionIdAllowed}/candidates-import-sheet`,
-        payload: {},
-        headers: { authorization: generateValidRequestAuthorizationHeader(user.id) },
-      };
-      // when
-      const promise = server.inject(options);
-
-      // then
-      return promise.then((response) => {
-        expect(response.statusCode).to.equal(200);
-      });
-    });
-  });
-});
